Update documentation and execution (minor changes); improve computational efficiency and ecological validity
While the simulations associated with the first submission worked as intended, I realized there was room for a non-trivial improvement. Specifically, the feedback function needed to be bounded between -1 and 1 in order to be used in the learning rule that updates feedback values. My initial approach was to truncate values exceeding these limits, forcing them to the minimum or maximum acceptable values, respectively. However, this dampened the variability expected in organic systems. To address this, I implemented a scaling-by-the-maximum rule for the gains that constitute the feedback path, applying it before computing the feedback increment score. This approach constrained gain values within the -1 to 1 range thus naturally preserving variability, ensuring that the target metric displayed its expected fluctuations. This improvement was integrated into the corresponding scripts, and Figures 2 and 3 were updated accordingly to reflect the new data. Notably, these modifications did not alter the main outcome.

I also identified that in the main iterative routine a section of code recording data was unnecessarily repeating with every loop, despite only the final iteration’s data being preserved. By enclosing this part within an if statement that executes only during the final loop (i = loops), I was able to achieve sometimes a 23% reduction in execution time by eliminating this repetitive recording and deletion of the same data frame. 

In addition, at the beginning of each script I added an explanation on how it builds upon the previous one to tackle duplication. Therefore, the expectation of duplication is nested in readers’ mind before they encounter the resemblances in the code. This creates a context where the duplication is expected and serves a pedagogical purpose rather than being an oversight. Minor duplications were tracked and fixed across the scripts.